requirements.txt
# Fall Detection System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core ML & Computer Vision
ultralytics>=8.0.0        # YOLOv8 - fall detection model
opencv-python>=4.8.0      # Video capture and processing
numpy>=1.24.0             # Array operations

# Firebase
firebase-admin>=6.2.0     # Firebase SDK (Realtime DB + FCM)

# Utilities
python-dotenv>=1.0.0      # Load .env config file
Pillow>=10.0.0            # Image handling
requests>=2.31.0          # HTTP requests

# Optional: SMS alerts via Twilio
# twilio>=8.0.0

# Optional: GPU acceleration (uncomment if you have NVIDIA GPU)
# torch>=2.0.0
# torchvision>=0.15.0
